+++ lib/synheart_core.dart
@@ -123,6 +123,7 @@ export 'src/modules/capabilities/capability_token.dart';
 export 'src/modules/consent/consent_module.dart';
 export 'src/modules/consent/consent_effective_state.dart';
 export 'src/modules/consent/consent_form.dart';
+export 'src/modules/consent/consent_type_meta.dart';
 export 'src/modules/consent/consent_profile.dart';
 export 'src/modules/consent/consent_token.dart';
 export 'src/modules/consent/consent_ui.dart';